move convertParamsErrorsToAudits to common
this function is needed in frontend to show errors in the validation one pager
