+++ packages/evolution-backend/src/services/audits/AuditUtils.ts
@@ -4,28 +4,7 @@
  * This file is licensed under the MIT License.
  * License text available at https://opensource.org/licenses/MIT
  */
-import { AuditForObject, Audits } from 'evolution-common/lib/services/audits/types';
-import slugify from 'slugify';
-
-export const convertParamsErrorsToAudits = (
-    errors: Error[],
-    objectData: Pick<AuditForObject, 'objectType' | 'objectUuid'>
-): AuditForObject[] => {
-    const audits: AuditForObject[] = [];
-    for (let i = 0, countI = errors.length; i < countI; i++) {
-        const error = errors[i];
-        const errorCode = slugify(error.message);
-        audits.push({
-            version: 1,
-            errorCode,
-            message: error.message,
-            isWarning: false,
-            ignore: false, // params errors should never be ignored
-            ...objectData
-        });
-    }
-    return audits;
-};
+import { Audits } from 'evolution-common/lib/services/audits/types';
 
 // Will ignore audits set to ignore except if overrideIgnoredAudits is true.
 // Old existing audits not triggering errors in new version will be removed.
