+++ packages/evolution-backend/src/services/audits/__tests__/AuditUtils.test.ts
@@ -6,37 +6,8 @@
  */
 import { v4 as uuidV4 } from 'uuid';
 
-import { mergeWithExisting, convertParamsErrorsToAudits } from '../AuditUtils';
-import slugify from 'slugify';
+import { mergeWithExisting } from '../AuditUtils';
 import { Audits } from 'evolution-common/lib/services/audits/types';
-
-const arbitraryUuid = uuidV4();
-
-describe('convertParamsErrorsToAudits', () => {
-    it('should convert error messages to audits', async () => {
-        const errors = [new Error('Error message 1'), new Error('Error message 2')];
-        const result = convertParamsErrorsToAudits(errors, { objectType: 'interview', objectUuid: arbitraryUuid });
-
-        expect(result[0]).toEqual({
-            errorCode: slugify('Error message 1'),
-            objectUuid: arbitraryUuid,
-            objectType: 'interview',
-            message: errors[0].message,
-            version: 1,
-            ignore: false,
-            isWarning: false
-        });
-        expect(result[1]).toEqual({
-            errorCode: slugify('Error message 2'),
-            objectUuid: arbitraryUuid,
-            objectType: 'interview',
-            message: errors[1].message,
-            version: 1,
-            ignore: false,
-            isWarning: false
-        });
-    });
-});
 
 describe('mergeWithExisting', () => {
     it('should merge new audits with existing audits', async () => {
