+++ packages/evolution-common/package.json
@@ -26,6 +26,7 @@
     "i18next": "^22.4.15",
     "libphonenumber-js": "^1.10.47",
     "lodash": "^4.17.21",
+    "slugify": "^1.6.6",
     "typescript": "^4.9.4",
     "uuid": "^8.3.2"
   },
